♻️ Move Default function to options.go
